Improve interface of LRUCache, ResultTable
LRUCache now uses a tryEmplace method which only constructs a new element if
non existed for the key beforehand. This is made race-free by returning a
read-only reference to any item already associated with the key if no item was
created. We need to do this in one operation since otherwise an item in the
cache may already have been deleted between the time we figure out that it
already existed and the time we try to access it.

In ResultTable we allow isFinished() and awaitFinished() to be called on const
references as only the state of the mutex/condition_variable is modified not
the object itself.