Don't force "en_US.utf8" locale
Instead of using setlocale(LC_CTYPE, "en_US.utf8") and thus forcing "en_US" we
now use setlocale(LC_CTYPE, "") which configures the program to use the default
locale's character set instead of forcing the uber-portable "C" or "POSIX"
locale.  (see "man 3 setlocale").

On a sane system this will use UTF-8, we do not care about insane systems.

The main advantage of this change is that it allows to use docker containers
without running "locale-gen" in the container thus making the image smaller.