Increase QET cache and STXXL memory size
We're saving a lot of RAM with more efficient storage of the OPS/OSP
meta data and soon with vocabulary improvements. So we should put some
of that saved memory to good use.

Thus this commit increases the size of the query execution tree (QET)
cache ten fold. It also doubles the amount of memory STXXL may use.

When testing with Aqqu with a pure QLever backend this increases query
performance by at least an order of magnitude with minimal increase im
RAM use.