Don't estimate OPTIONAL_JOIN size as 0 fixes #108
When the OPTIONAL_JOIN has a an empty side it would estimate it's result
as empty. For estimates of size zero JOIN would then take the estimate
as exact and stop early.

The current fix is just incrementing the size estimate by one to make
sure it's never 0. In the future we may want to get an exact estimate
for 0.

While we are here also remove redundant `getRootOperation()` calls in
Join.cpp (`getResult()` already uses the `_rootOperation`) and add some
useful `LOG(TRACE)`