Updated MetaDataConverter to handle basically any old format
-- from every format for IndexMetaData we have introduced so far it is
now possible to automatically determine the correct version to read and
convert it

Eliminated -l flag for ServerMain

- this boolean flag can not be chosen by the user and has to match the
  settings chosen at index-build time. This information is passed via
  the .meta-data.json file

Prefix Compression now also takes into account that
we ALWAYS add a code prefix even if there's nothing to compress
that way, the Prefix " will also be compressed (one byte of saving per
literal that is not being compressed otherwise)

Separation between different types of vocabulary.

We have disabled externalizing parts of an uncompressed vocabulary,
because there the method resolving IDs is returning a reference which
also does not work with external literals.