Use QueryExecutionContext to track cache pinning
This allows to pin all subtrees of a query to the cache and have that as
a per query option. On the QLever server this is exposed as the
"pinsubtrees" URL param. If set to true all subtrees in the query will
be cached. This is especially useful for completion queries and the
like.