BIND with more general binary operations
Now allows the four basic arithmetic binary operations:  + - * /

TODO: Since the - is not recognized by a SYMBOL, the minus is currently
realized via | (like a minus but vertical). I hope that Johannes finds a
better way.

I tried it out on a few example queries and it worked.

TODO: If one of the two operands is a constant (from a previous BIND,
like BIND(100 AS ?hundred), the operations don't work as expected and I
am not sure why. Maybe Johannes can help.