Introduce own class for QLever's internal entity IDs (#637)
So far, a lot of different things were an "ID" in QLever: Qlever's internal entity IDs, indices of various vocabularies, offsets, etc. This PR is a first step towards a cleaner typing for these IDs, by introducing an own class "Id" for Qlever's internal entity IDs. In particular, we need this for encoding values from various data types into an ID (one of the next PRs).