Integrate Value IDs (#648)
The value IDs are now in effect! Doubles and ints are now stored as part of the ID (and not, as before, as a special :v:... string), their order is 100% correct now, and different value types can be mixed now in one column. Other types still use the :v:... strings (like dates) or are stored as string literals and not values, but this will be easy to fix now (especially since we already have a Date class). Efficiency-wise, there is still room for improvement, in particular, in the typical case where all bindings of a variable have the same type.