Enable parallel execution of tests (#672)
So far, parallel execution of our tests failed because of shared file names, etc. Now filenames are as distinct as possible and test cases that must not be executed in parallel with any other tests (these are currently the tests involving STXLL, which share a common STXLL disk file) are explicitly marked as such.