Fix the "missing KB name" bug and refactor URL parameter handling (#691)
1. For a long time now, the option -K from the command line for setting the
index name had no effect. This is now fixed.
2. Also allow a URL parameter "index-description" for setting this name
for a running server, and an analogous parameter "text-description" for
the text index.
3. A subset of the URL parameters now requires that the new URL
parameter "access-token" is set to the "right" value. What is "right" is
currently hard-coded. In a PR to come, the "right" value should be set
via a command-line argument when starting the server.
4. Make icu::CaseMap::utf8ToLower work with std::string_view (leftover
from PR #659).