Bugfix: catch exceptions that occur in the pattern trick computation (#727)
So far, exceptions thrown in the pattern trick computation crashed the server. This is now fixed. Also, ql:has-predicate for an ID that is not a vocabulary index (for example, a value) no longer throws an exception, but such IDs are simply ignored (correctly because they have no predicates). Added two queries in the end-to-end test.