Fix parser error messages that contain multibyte UTF-8 characters (#808)
Fixes #801 
ANTLR start and stop index for tokens refer to unicode codepoints, not to bytes, so they have to first be converted to bytes before applying them to the UTF-8 encoded input to the parser when formatting a parseError. This is now fixed, including the false assertion that was previously thrown for queries that contained multibyte characters.