Fix the computation of the variable column map for UNION (#809)
Merging the variable column map of the two sides of a UNION becomes non-trivial when variables on either side can be deleted because they are not visible outside of a subquery (see PR 805). The map is now computed correctly in all cases.

On the side, now use the address sanitizer (option -fsanitize=address) for our continuous native build. The address sanitizer has functionality similar to Valgrind, but with a much smaller slowdown when executing the code. It helped us find this nasty bug.