Improve size estimate for has-predicate scan with context (#807)
The previous size estimate for a has-predicate scan with context gave the same estimate as a full has-predicate scan, although the latter is much more expensive. Through some miracle, QLever always picked the better query plan in the past, but after PR #795 it no longer did. With this change, the better plan in this situation is now indeed strictly cheaper. Apart from that, the cost estimates are still simplistic, but that is work for another PR.