Support (almost) arbitrary expressions for FILTER and HAVING (#810)
This completes our transition to parsing the complete SPARQL query using our new ANTLR-based parser. As a side-effect, a lot of code from the old parser, which was still needed for parsing FILTER expressions, is now gone. The SparqlParser class is still there, but all it does now is provide a method parseQuery that calls the new ANTLR-based parser.

On the side, fixes https://github.com/ad-freiburg/qlever/issues/800 . Also fixes the longer-standing bug that so far negation was a no-op.