Use type Variable instead of std::string inside TripleComponent (#813)
This is another step towards consistently using "strong" types. We still use std::string to store variable names in other places of the code, notably the map from variable names to column indexes. Also, literals and IRIs are still both stored as std::string, using the first character to differentiate (" for literals, < for IRIS). This is work for future PRs.