Use local vocabulary for VALUES clause (#822)
So far, a row of a VALUES clause was ignored when one of the values did not occur in the normal vocabulary and could not be "folded" into an ID (like an integer or a double). Now all rows are considered, and OOV values are added to the local vocabulary.

On the side, the SPARQL parser for the VALUES clause is modified so that it does not produce an `std::string` for each value, like it did so far, but a `TripleComponent`. That change led to further refactoring and simplification of the surrounding code.