Efficient COUNT for certain queries involving a full index scan (#830)
The following kind of query is now also processed more efficiently, avoiding an actual scan of the whole index. This continues work from #824.

SELECT ?p (COUNT(?p) AS ?count) WHERE { ?s ?p ?o } GROUP BY ?p

The GROUP BY variable (which then also has to be the variable in the SELECT clause) can be by any of ?s, ?p, or ?o. The argument of the COUNT can be any of the three variables (the result is the same). Either the variable itself or the count alias can be omitted in the SELECT clause. For GROUP BY ?p, the query is very fast because the metadata for the Pxx permutations is held in RAM. For GROUP BY ?s and GROUP BY ?o, there is one disk access per subject or object, respectively.