Efficient and correct COUNT of total number of subjects, predicates, or objects (#836)
The following kind of query is now processed more efficiently, avoiding an actual scan of the whole index. This continues work on the efficient processing of typical statistics queries from #824 and #830.

SELECT (COUNT(DISTINCT ?s) AS ?count) WHERE { ?s ?p ?o }

The argument of the COUNT DISTINCT can be either ?s or ?p or ?o. The respective counts are precomputed and stored as part of the index meta data. They are shown under "Index Information" in the QLever UI. The triples QLever adds internally are NOT considered anymore in these counts (they used to be), but there are separate counts for those in the index meta data, too.

IMPORTANT: This requires rebuilding the index (or manually adjusting the .meta-data.json file from the index build). It also requires a code change in the QLever UI because the names in the JSON object returned by QLever for cmd=stats have changed.