Column-major order for the ID tables (#841)
The `IdTable` is QLever's central data structure, which stores intermediate as well as final results in ID space. This PR changes the layout of these tables from row-major to column-major.

The interface to the class is largely preserved, so that the code using the class can more or less remain as it is. However, several important improvements and some small changes were made, most notably: (1) it should now be harder to use the class wrongly; (2) better names for some of the methods; (3) an explicit `clone()` function when a deep copy happens; (4) several convenience functions like == between two tables.