More accurate and informative runtime information (#842)
The runtime information now provides the following additional information: which operations were executed during query planning, the time needed for query planning (without the time for the aforementioned operations), which operations were cached and not pinned, and which operations were cached and pinned.

NOTE: The JSON with the runtime info returned by QLever now looks a little different. This requires a corresponding update of the QLever UI code.