When resizing `IdTable`, ensure minimum growth by factor 1.5 (#848)
A `resize` operation now increases the capacity by at least the growth factor. The amortized cost of each of a sequence of "small" `resize` operations is therefore constant. A growth factor of 1.5 was also present in the old row-based `IdTable`. It is considered to be better than a factor of 2 also by many industry experts (see for example `folly::fbvector`).