Add some built-in functions (#766)
Add the following built-in functions: STR, STRLEN, DAY, MONTH, YEAR, RAND.  Add unit tests both for the parsing and the function evaluation. The implementations of the functions are not yet efficient (except for RAND, which uses our own `FastRandomIntGenerator`), but they work and we now have a few examples that will be helpful when adding more such functions in the near future.