Fix the initialization of `FastRandomIntGenerator` (#887)
The `FastRandomIntGenerator` works with four 64-bit ints internally. These four ints were initialized with 32-bit seeds. The effect was that the first number was not really random (it was a small random number). Now the four ints are initialized with 64-bit seeds and the very first number is random, too.