Correct implementation of ORDER BY (#874)
The `ORDER BY` operation now computes the correct ordering in all cases. The internal operations `OrderBy` and `Sort` now do different things. `OrderBy` sorts as expected for an `ORDER BY` (this requires a non-trivial comparison between values). `Sort` sorts by internal IDs (which allows fast comparisons and is good enough for merge JOINs), and can be explicitly invoked by using `INTERNAL SORT BY` instead of `ORDER BY`.

IMPORTANT: As a consequence, two warmup queries of our AC (autocompletion) queries in the Qlever UI now have to be rewritten to use `INTERNAL SORT BY` instead of `ORDER BY`. 

Also, there are now two comparators, which differ in how they handle a comparison of two values with different datatypes (in particular, when one value is undefined and the other is not). For one comparator, a `<`, `<=`, `=`, `>=`, `>` comparison is always false when the datatypes are different (for example, used by `FILTER (?x < 42)`, which should filter out all undefined values). For the other comparator, the result depends on a deterministic order of the datatypes (for example, used in an `ORDER BY`, where all undefined values now come at the beginning, for `ASC`, or end, for `DESC`).