Refactor code for exporting query results (#843)
Refactor the code that translates an internal query result (represented using IDs) to QLever JSON, SPARQL JSON, CSV, TSV, Turtle, or QLever's binary format. There was previously significant code duplication, inconsistencies in the translation process, and most of the code was in the `QueryExecutionTree` class, where it does not belong. The refactored code is now in a class `ExportQueryExecutionTree` and has extensive unit tests.