Fix comparisons involving UNDEF or incompatible types (#895)
`MIN` and `MAX` are now computed correctly using the same comparator as for `ORDER BY` (see #874), except that `UNDEF` values are excluded. For example, for `VALUES ?x { 5 4 "three" UNDEF }`, now `MIN(?x)` is `4` and `MAX(?x)` is `"three"` (because the string datatype comes after the number datatype in our arbitrary fixed ordering of datatypes). This fixes #857 .

Also, `!=` for two values with incompatible datatypes (for example, `3 != "three"`) is now always false, as required by the standard. For the other comparators (`<`, `<=`, `=`, `>=`, `>`) such comparisons were already false.

TODO: For `BIND`, the semantics of `!=` is still not 100% correct. For example, `BIND(3 != "three" AS ?x)` should leave `?x` unbound (`UNDEF`), but with the current code binds if to `false`.