Fix the docker build on Github actions. (#898)
The docker workflow used to run on `ubuntu-latest` but internally requires `Ubuntu 20.04`, because some packages are manually downloaded for exactly this version of Ubuntu. By explicitly requesting `Ubuntu 20.04` as the underlying OS for the Docker workflow on GitHub actions this workflow now also runs correctly when the `ubuntu-latest` tag changes to a newer version.

This eliminates another source of possible spurious failures of CI checks.