Basic support of SERVICE clause (#793)
Support SERVICE queries to a remote endpoint specified via an IRI. The query is sent using a POST request with headers `Content-Type: application/sparql-query` and `Accept: text/tab-separated-values`. The result TSV is parsed and words that are not contained in the index vocabulary are added to the local vocabulary of the result of the SERVICE operation.

Shortcomings of this implementation are: there is no timeout, the TSV is first read (and stored) completely and only then processed further, TSV is simple to parse but may be ambiguous with respect to how IRIs or literals are encoded, the `SILENT` keyword (telling the engine to ignore the SERVICE clause if the query fails) is currently not supported, variable endpoints are currently not supported. Also, the query is currently sent after query planning, so the query planner has no information about the size, cost, or multiplicities of the result.