Make `IdTable` use a vector<vector<Id>> to store the columns (#891)
The previous implementation used one `vector<Id>` for all columns of an `IdTable`, to make row-wise iteration simpler. Now we use one separate `vector<Id>` per column. This does not make row-wise iteration slower, but makes the code significantly easier in several places and enables various optimizations in the future (for example, generating views of a subset of the columns or dropping columns that are no longer needed).