Enable static analysis via SonarCloud
Add a GitHub action that runs the static
analysis from SonarCloud and uploads it
to the SonarCloud servers. This action is run for every PR and for every push to the Master branch.

Note: A complete scan of the QLever code base currently takes 2.5 hours. Maybe this will be faster for small PRs because of SonarCloud's cache for unchanged translation units.