More correct implementation of AS (#897)
The implementation of `AS` so far was not correct in several corner cases and the error messages were not optimal. Both the correctness and the error messages are significantly improved now. In particular, this fixes #856.

IMPORTANT: An alias like `(SAMPLE(?profession) AS ?profession)` in the `SELECT` clause was allowed so far by QLever, but is not allowed by the SPARQL standard (for good reasons because it can lead to ambiguity when the variable is used again in a subsequent alias). It is now no longer allowed. This may require rewriting some of our example queries in the QLever UI.