Fix several bugs and code smells reported by SonarCloud (#906)
In particular, improve our handling of destructors that can throw an exception. When possible, this does not cause the program to terminate any longer. When a termination is unavoidable because there is no way to recover from the error (for example, when a destructor writes finalizing information to a file, but the file can not be written to anymore), at least terminate with a proper error message.