Support all expressions in HAVING and ORDER BY (#914)
So far, certain expressions in HAVING or ORDER BY were not supported. For example, it was not possible to use a variable defined as the target of an alias (in the SELECT clause) in a HAVING or ORDER BY expression. In combination with GROUP BY, it was not possible to use aggregates of the non-grouped variables. Now all valid expressions are supported.

On the side, introduce a macro `AD_EXPECT_THROW_WITH_MESSAGE` that allows unit tests that check not only that an error occured, but also that it occurred with the right error message.