Refactor the `ResultTable` class and how we use it (#925)
All our `Operation`s used to have a `ResultTable*` as argument and were called with a pointer to an empty `ResultTable`. This was ugly and error-prone. Now each `Operation` returns a `ResultTable`, which is constructed by the operation. After the operation, the `ResultTable` as well as its `LocalVocab` are `const` and coherent (for example, the information by which columns the table is sorted is in sync with the contents of the table). This makes the code much safer and cleaner. In particular, we don't need the hack anymore to make the `LocalVocab` read-only after an initial construction phase.