More efficient column-based UNION (#935)
The previous code computed the result of the UNION row by row, copying the values from each row (first from the left, then from the right operand) to the correct position in the result (the left and right operand may have different sets of variables). Now whole columns from the left and right operand are copied to the result (or parts of the result filled with UNDEF if the respective operand is missing the respective variable). The copying and filling happens in chunks so that we can have timeout checks inbetween the chunks. On two example queries on Wikidata, the new code was twice as fast.