Fix subtle bug in variable to columns map for JOIN (#938)
There was a subtle bug in the map from variables to columns in the JOIN operation, namely when the operands contained "invisible" columns (that is, columns that were computed in previous operations but then selected out). In particular, this bug showed in one of the warmup AC queries from the QLever UI. That bug is now fixed and corresponding unit tests and a query in our end-to-end test were added.