Fix issues marked by SonarCloud as "bug" (#943)
Most of these issues were of the following kind:
When a class `S` has a templated constructor (e.g. `S(auto&& arg)`), then this constructor will often wrongly be preferred to the copy constructor. E.g. when trying to copy construct from a non-const `S`, the actual signature is `S(S&)` for which the templated constructor will be a better match than the copy constructor (`S(const S&)`). A fix for this subtle issue is to constrain the templated constructor to only take arguments of types that are sufficiently different from `S`.